+++ timesketch/ui/graph/cypher-query.component.ts
