+++ timesketch/ui/graph/cypher-query.data.ts
