+++ timesketch/ui/graph/main.component.ts
