+++ timesketch/ui/graph/models.ts
