+++ timesketch/ui/graph/sidebar.component.html
