+++ timesketch/ui/graph/sidebar.component.ts
